requirements.txt
# ── Web Framework ────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.36
sqlalchemy[asyncio]
greenlet
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# ── Auth ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1

# ── AI / ML ───────────────────────────────────────────────
anthropic>=0.84.0
google-genai>=1.0.0
faster-whisper==1.2.1
mediapipe==0.10.32
deepface==0.0.99
opencv-python-headless==4.13.0.92  # Use headless if on a server
numpy>=2.0.0
scipy>=1.17.1

# ── PDF / Document Parsing ────────────────────────────────
pdfplumber==0.11.4
python-docx==1.2.0

# ── Storage / Cloud ───────────────────────────────────────
boto3==1.35.80

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0